Integrate message tracking with entity #1208

+++ clustered/server/src/main/java/org/ehcache/clustered/server/EhcacheActiveEntity.java
@@ -25,6 +25,7 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import org.ehcache.clustered.common.Consistency;
@@ -33,6 +34,7 @@ import org.ehcache.clustered.common.internal.ClusteredEhcacheIdentity;
 import org.ehcache.clustered.common.PoolAllocation;
 import org.ehcache.clustered.common.internal.exceptions.ClusterException;
 import org.ehcache.clustered.common.internal.exceptions.IllegalMessageException;
+import org.ehcache.clustered.common.internal.exceptions.InvalidClientIdException;
 import org.ehcache.clustered.common.internal.exceptions.InvalidOperationException;
 import org.ehcache.clustered.common.internal.exceptions.InvalidStoreException;
 import org.ehcache.clustered.common.internal.exceptions.LifecycleException;
@@ -94,6 +96,7 @@ class EhcacheActiveEntity implements ActiveServerEntity<EhcacheEntityMessage, Eh
   private final ConcurrentHashMap<String, Set<ClientDescriptor>> storeClientMap =
       new ConcurrentHashMap<String, Set<ClientDescriptor>>();
 
+  private final ConcurrentHashMap<ClientDescriptor, UUID> clientIdMap = new ConcurrentHashMap<>();
   private final ReconnectDataCodec reconnectDataCodec = new ReconnectDataCodec();
   private final ServerStoreCompatibility storeCompatibility = new ServerStoreCompatibility();
   private final EhcacheEntityResponseFactory responseFactory;
@@ -259,6 +262,7 @@ class EhcacheActiveEntity implements ActiveServerEntity<EhcacheEntityMessage, Eh
     }
     clientState.attach();
     ReconnectData reconnectData = reconnectDataCodec.decode(extendedReconnectData);
+    clientIdMap.put(clientDescriptor, reconnectData.getClientId());
     Set<String> cacheIds = reconnectData.getAllCaches();
     for (final String cacheId : cacheIds) {
       ServerStoreImpl serverStore = ehcacheStateService.getStore(cacheId);
@@ -537,7 +541,9 @@ class EhcacheActiveEntity implements ActiveServerEntity<EhcacheEntityMessage, Eh
    */
   private void configure(ClientDescriptor clientDescriptor, ConfigureStoreManager message) throws ClusterException {
     validateClientConnected(clientDescriptor);
-    ehcacheStateService.configure(message);
+    if (ehcacheStateService.getClientMessageTracker().shouldConfigure(message.getClientId(), message.getId())) {
+      ehcacheStateService.configure(message);
+    }
     this.clientStateMap.get(clientDescriptor).attach();
   }
 
@@ -551,6 +557,10 @@ class EhcacheActiveEntity implements ActiveServerEntity<EhcacheEntityMessage, Eh
    */
   private void validate(ClientDescriptor clientDescriptor, ValidateStoreManager message) throws ClusterException {
     validateClientConnected(clientDescriptor);
+    if (clientIdMap.get(clientDescriptor) != null && clientIdMap.get(clientDescriptor).equals(message.getClientId())) {
+      throw new InvalidClientIdException("Client ID : " + message.getClientId() + " is already being tracked by Active");
+    }
+    clientIdMap.put(clientDescriptor, message.getClientId());
     ehcacheStateService.validate(message);
     this.clientStateMap.get(clientDescriptor).attach();
   }
@@ -578,14 +588,20 @@ class EhcacheActiveEntity implements ActiveServerEntity<EhcacheEntityMessage, Eh
     if(createServerStore.getStoreConfiguration().getPoolAllocation() instanceof PoolAllocation.Unknown) {
       throw new LifecycleException("Clustered tier can't be created with an Unknown resource pool");
     }
-
+    boolean isDuplicate = isLifeCycleMessageDuplicate(createServerStore);
     final String name = createServerStore.getName();    // client cache identifier/name
+    ServerStoreImpl serverStore;
+    if (!isDuplicate) {
 
-    LOGGER.info("Client {} creating new clustered tier '{}'", clientDescriptor, name);
+      LOGGER.info("Client {} creating new clustered tier '{}'", clientDescriptor, name);
 
-    ServerStoreConfiguration storeConfiguration = createServerStore.getStoreConfiguration();
+      ServerStoreConfiguration storeConfiguration = createServerStore.getStoreConfiguration();
 
-    ServerStoreImpl serverStore = ehcacheStateService.createStore(name, storeConfiguration);
+      serverStore = ehcacheStateService.createStore(name, storeConfiguration);
+    } else {
+      serverStore = ehcacheStateService.getStore(name);
+    }
+
     serverStore.setEvictionListener(new ServerStoreEvictionListener() {
       @Override
       public void onEviction(long key) {
@@ -668,9 +684,18 @@ class EhcacheActiveEntity implements ActiveServerEntity<EhcacheEntityMessage, Eh
       throw new ResourceBusyException("Cannot destroy clustered tier '" + name + "': in use by " + clients.size() + " other client(s)");
     }
 
-    LOGGER.info("Client {} destroying clustered tier '{}'", clientDescriptor, name);
-    ehcacheStateService.destroyServerStore(name);
+    boolean isDuplicate = isLifeCycleMessageDuplicate(destroyServerStore);
+
+    if (!isDuplicate) {
+      LOGGER.info("Client {} destroying clustered tier '{}'", clientDescriptor, name);
+      ehcacheStateService.destroyServerStore(name);
+    }
+
     storeClientMap.remove(name);
+  }
+
+  private boolean isLifeCycleMessageDuplicate(LifecycleMessage message) {
+    return ehcacheStateService.getClientMessageTracker().isDuplicate(message.getId(), message.getClientId());
   }
 
   /**
